Enum signed : RFC compliance

+++ src/lib/http/HttpClient.cpp
@@ -12,7 +12,7 @@
 using asio::ip::tcp;
 
 int
-http_request(std::string domain, std::string path, int port, std::string& ret)
+http_request(std::string domain, std::string path, unsigned short port, std::string& ret)
 {
     try
     {
